Thu, Nov 9, 2023, 6:18 PM -07:00

+++ 1-lecture/calculator.c
@@ -6,7 +6,7 @@ int main(void)
     long x = get_long("x: ");
     long y = get_long("y: ");
 
-    float z = (float) x / (float) y;
+    double z = (double) x / (double) y;
 
     printf("%.20f\n", z);
 }
